Deleted Duke, DukeTest. Changed build files. Packaged the java files.
