--- src/main/java/Command.java
@@ -1,9 +1,0 @@
-
-
-public abstract class Command {
-    public abstract void execute(WatchedList watchedList, ToWatchList toWatchList, Ui ui, Storage storage);
-}
-
-
-
-
