--- src/main/java/ListCommand.java
@@ -1,10 +1,0 @@
-
-public class ListCommand extends Command {
-
-    @Override
-    public void execute(WatchedList watchedList, ToWatchList toWatchList, Ui ui, Storage storage) {
-        // list the watched list
-        // watchedList.list();
-        Ui.showListMessage(watchedList);
-    }
-}
