--- src/main/java/ViewReviewCommand.java
@@ -1,8 +1,0 @@
-
-public class ViewReviewCommand extends Command {
-
-    @Override
-    public void execute(WatchedList watchedList, ToWatchList toWatchList, Ui ui, Storage storage) {
-        watchedList.viewReview(ui, watchedList);
-    }
-}
