+++ src/main/java/seedu/moviemate/command/AddReviewCommand.java
@@ -1,0 +1,16 @@
+package seedu.moviemate.command;
+
+import seedu.moviemate.storage.Storage;
+import seedu.moviemate.movie.ToWatchList;
+import seedu.moviemate.ui.Ui;
+import seedu.moviemate.movie.WatchedList;
+
+public class AddReviewCommand extends Command {
+
+    //maybe can do some parse to get the index from String here later.
+    @Override
+    public void execute(WatchedList watchedList, ToWatchList toWatchList, Ui ui, Storage storage) {
+        watchedList.setReview(ui, watchedList);
+    }
+
+}
