+++ src/main/java/seedu/moviemate/command/ExitCommand.java
@@ -1,0 +1,18 @@
+package seedu.moviemate.command;
+
+import seedu.moviemate.storage.Storage;
+import seedu.moviemate.movie.ToWatchList;
+import seedu.moviemate.ui.Ui;
+import seedu.moviemate.movie.WatchedList;
+
+public class ExitCommand extends Command {
+
+    @Override
+    public void execute(WatchedList watchedList, ToWatchList toWatchList, Ui ui, Storage storage) {
+        //Ui...
+    }
+
+    public static boolean isExit(Command command) {
+        return command instanceof ExitCommand; // instanceof returns false if it is null
+    }
+}
