+++ src/main/java/seedu/moviemate/movie/ToWatchList.java
@@ -1,0 +1,19 @@
+package seedu.moviemate.movie;
+
+import java.util.ArrayList;
+
+/**
+ * A list of movies the user plans to watch.
+ */
+public class ToWatchList extends MovieList{
+
+    public ToWatchList() {
+        super();
+    }
+
+    public ToWatchList(ArrayList<String[]> load) {
+        super(load);
+    }
+
+    // TODO: more methods here specific to toWatchList.
+}
