+++ text-ui-test/runtest.bat
@@ -1,3 +1,5 @@
+EXIT /B
+
 @echo off
 setlocal enableextensions
 pushd %~dp0
